% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPI_module.R
\name{KPI_infobox_Server}
\alias{KPI_infobox_Server}
\title{Server function of module to render SORMAS-Stats Specific KPI valueBoxes}
\usage{
KPI_infobox_Server(id, data, panel)
}
\arguments{
\item{data}{reactive data you want to display (casedata,eventdata or contactdata)}

\item{panel}{set to which type of KPI's you want to render ("contact","case","event"), defines ifelse logic}
}
\value{
A list of shiny.tag's containing valueBoxes for SORMAS-Stats
}
\description{
Server function of module to render SORMAS-Stats Specific KPI valueBoxes
}
