% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPI_module.R
\name{KPI_infobox_UI}
\alias{KPI_infobox_UI}
\title{UI function of module to render SORMAS-Stats Specific KPI valueBoxes}
\usage{
KPI_infobox_UI(id)
}
\arguments{
\item{id}{Namespace id}
}
\value{
A list of rendered HTML div’s of valueBoxes for SORMAS-Stats
}
\description{
UI function of module to render SORMAS-Stats Specific KPI valueBoxes
}
