% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendervalueBoxfromList.R
\name{rendervalueBoxfromList}
\alias{rendervalueBoxfromList}
\title{render a list of shiny columns containing valueBoxes}
\usage{
rendervalueBoxfromList(list, icon, color)
}
\arguments{
\item{list}{List of named lists containing the values for the boxes}

\item{icon}{An <i> (icon) HTML tag.}

\item{color}{A color for the box. Valid colors are listed in validColors. help/library/shinydashboard/help/validColors}
}
\value{
A list of shiny.tag's containing valueBoxes
}
\description{
render a list of shiny columns containing valueBoxes
}
