% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testapp.R
\docType{data}
\name{Testapp}
\alias{Testapp}
\alias{ui}
\title{Testapp to show functionality}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
ui
}
\description{
Testapp to show functionality
}
\keyword{datasets}
